test(inference): split Hermes smoke verifier spec
