+++ src/lib/inference/onboard-probes.test.ts
@@ -5,7 +5,6 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
-import { runVerifyOnboardSmokeHarness } from "../../../test/helpers/onboard-smoke-verifier-harness";
 
 const {
   getChatCompletionsProbeCurlArgs,
@@ -18,42 +17,9 @@ const {
   isSandboxInternalUrl,
   probeOpenAiLikeEndpoint,
   RETRIABLE_HTTP_PROBE_STATUSES,
-  shouldSmokeOpenAiLikeOnboardRoute,
 } = require("../../../dist/lib/inference/onboard-probes");
 
 describe("OpenAI-compatible inference probe response parsing", () => {
-  it("does not host-smoke Hermes Provider with the ambient OPENAI_API_KEY", () => {
-    expect(shouldSmokeOpenAiLikeOnboardRoute("hermes-provider", "OPENAI_API_KEY")).toBe(false);
-    expect(shouldSmokeOpenAiLikeOnboardRoute("hermes-provider", "NOUS_API_KEY")).toBe(true);
-    expect(shouldSmokeOpenAiLikeOnboardRoute("openai-api")).toBe(true);
-  });
-
-  it("skips only the Hermes OAuth smoke path in the runtime verifier", () => {
-    const calls = runVerifyOnboardSmokeHarness([
-      { credentialEnv: "OPENAI_API_KEY" },
-      { credentialEnv: "NOUS_API_KEY" },
-      { credentialEnv: "OPENAI_API_KEY", forceOpenAiLike: true },
-    ]);
-    expect(
-      calls.filter((call) =>
-        ["resolveProviderCredential", "getCredential", "runCurlProbe"].includes(call[0]),
-      ),
-    ).toEqual([
-      ["resolveProviderCredential", "NOUS_API_KEY"],
-      [
-        "runCurlProbe",
-        "https://api.example.com/v1/chat/completions",
-        "Authorization: Bearer resolved-NOUS_API_KEY",
-      ],
-      ["resolveProviderCredential", "OPENAI_API_KEY"],
-      [
-        "runCurlProbe",
-        "https://api.example.com/v1/chat/completions",
-        "Authorization: Bearer resolved-OPENAI_API_KEY",
-      ],
-    ]);
-  });
-
   it("detects tool-calling responses payloads conservatively", () => {
     expect(
       hasResponsesToolCall(
